auto_reply_bridge/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
# RAG vector store (optional — RAG is silently disabled if not installed)
chromadb==0.6.3
